LPD-56134 - Simplify
