+++ workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/i18n/en_US.ts
@@ -67,6 +67,7 @@ export default {
 	'approved': 'Approved',
 	'apps': 'Apps',
 	'area': 'Area',
+	'areas': 'Areas',
 	'available': 'Available',
 	'back': 'Back',
 	'back-to-apps': 'Back to Apps',
@@ -285,6 +286,7 @@ export default {
 	'license-prices': 'License Prices',
 	'license-type': 'License Type',
 	'licenses': 'Licenses',
+	'licensing': 'Licensing',
 	'liferay-learn': 'Liferay Learn',
 	'liferay-paas': 'Liferay PaaS',
 	'liferay-publisher-license-agreement':
@@ -364,6 +366,7 @@ export default {
 		'Please, review before submitting. Once sent, you will not be able to edit any information until this submission is completely reviewed by Liferay.',
 	'please-select-the-account-you-wish-to-link-to-your-liferay-dxp-below':
 		'Please select the account you wish to link to your Liferay DXP below',
+	'pricing': 'Pricing',
 	'privacy-policy': 'Privacy Policy',
 	'product': 'Product',
 	'product-unavailable': 'Product Unavailable',
@@ -457,6 +460,7 @@ export default {
 	'start-date': 'Start Date',
 	'start-trial': 'Start Trial',
 	'status': 'Status',
+	'storefront': 'Storefront',
 	'submit': 'Submit',
 	'submit-solution': 'Submit Solution',
 	'subscription-type': 'Subscription Type',
@@ -464,6 +468,7 @@ export default {
 	'subtotal-discount': 'Subtotal Discount',
 	'success': 'success',
 	'summary': 'Summary',
+	'support-and-help': 'Support & Help',
 	'support-email-address': 'Support Email Address',
 	'support-phone-number': 'Support Phone Number',
 	'support-url': 'Support URL',
