+++ workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/pages/PublisherDashboard/pages/NewAppFlow/PublishAppOutlet.scss
@@ -30,90 +30,11 @@
 	}
 }
 
-.card-section-body {
-	margin-top: 22px;
-}
-
-.card-section-body-section-tags {
+.submit-app-page-agreement {
 	display: flex;
-	flex-wrap: wrap;
-	gap: 8px;
-}
-
-.app-submit-section {
-	padding: 16px 0;
-}
-
-.submit-images-container img {
-	height: 64px;
-	width: 110px;
-}
-
-.submit-images-container .storefront-section-title {
-	font-size: 11px;
-	text-transform: capitalize;
-}
-
-.submit-images-container .video-player {
-	background-color: white;
-	width: 110px;
-	height: 64px;
-}
-
-.storefront-url-title {
-	font-size: 13px;
-	color: #282934;
-	font-weight: 600;
-}
-
-.storefront-description {
-	font-size: 13px;
-	color: #54555f;
-	font-weight: 400;
-}
-
-.important-note {
-	color: #54555f;
-	font-weight: 400;
-	font-size: 13px;
-	margin-top: 12px;
-}
-
-.card-section-body-description {
-	color: #54555f;
-	font-size: 16px;
-}
-
-.submit-logo-icon {
-	border-radius: 4px;
-	height: 48px;
-	width: 48px;
-}
-
-.submit-app-name {
-	font-weight: 700;
-	font-size: 24px;
-}
-
-.submit-app-version {
-	font-weight: 600;
-	font-size: 14px;
-	color: #54555f;
-}
-
-.unit-price-please {
-	font-size: 13px;
-	color: #54555f;
-}
-
-.licesing-price-type {
-	text-transform: capitalize;
-}
-
-.review-and-submit-app-page-agreement {
 	font-size: 13px;
 }
 
-.review-and-submit-app-page-agreement-highlight {
+.submit-app-page-agreement-highlight {
 	font-weight: 600;
 }
