+++ workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/pages/PublisherDashboard/pages/NewAppFlow/pages/Submit/Submit.scss
@@ -1,0 +1,120 @@
+.submit-app {
+	&-logo-icon {
+		border-radius: 4px;
+		height: 48px;
+		width: 48px;
+	}
+
+	&-name {
+		font-weight: 700;
+		font-size: 24px;
+	}
+
+	&-pricing {
+		&-description {
+			color: #54555f;
+			font-size: 13px;
+		}
+		&-icon {
+			color: #377cff;
+		}
+		&-title {
+			font-size: 18px;
+			font-weight: 600;
+		}
+	}
+
+	&-section {
+		padding: 16px 0;
+
+		.edit-button {
+			font-weight: 600;
+		}
+
+		.required-asterisk-icon {
+			height: 6px;
+			margin-bottom: 6px;
+			width: 6px;
+		}
+	}
+
+	&-section-body {
+		margin-top: 22px;
+
+		&-description {
+			color: #54555f;
+			font-size: 16px;
+			&-paragraph {
+				margin-bottom: 0;
+				margin-top: 22px;
+				word-break: break-word;
+				word-wrap: break-word;
+			}
+		}
+		&-tags {
+			display: flex;
+			flex-wrap: wrap;
+			gap: 8px;
+			margin-top: 22px;
+		}
+	}
+
+	&-storefront {
+		&-container {
+			img {
+				height: 64px;
+				width: 110px;
+			}
+			.icon-image {
+				color: #377cff;
+				height: 16px;
+				margin-top: 2px;
+				width: 16px;
+			}
+			.storefront-description {
+				color: #54555f;
+				font-size: 13px;
+				font-weight: 400;
+			}
+			.storefront-section-title {
+				font-size: 11px;
+				text-transform: capitalize;
+			}
+
+			.storefront-url-title {
+				font-size: 13px;
+				color: #282934;
+				font-weight: 600;
+			}
+			.video-player {
+				background-color: white;
+				width: 110px;
+				height: 64px;
+			}
+		}
+		&-important-note {
+			color: #54555f;
+			font-weight: 400;
+			font-size: 13px;
+			margin-top: 12px;
+		}
+	}
+
+	&-version {
+		font-weight: 600;
+		font-size: 14px;
+		color: #54555f;
+	}
+}
+
+.licensing-option-title {
+	font-size: 18px;
+	font-weight: 600;
+}
+.licesing-price-type {
+	text-transform: capitalize;
+}
+.licensing-unit-price {
+	font-size: 13px;
+	color: #54555f;
+}
